+++ packages/data/src/index.js
@@ -1,0 +1,6 @@
+import { Events } from 'marionette';
+
+export { default as Model } from './model.js';
+export { default as Collection } from './collection.js';
+export { DataApi, StateApi } from './api.js';
+export const triggerMethod = Events.triggerMethod;